+++ Sources/Layout/MessagesCollectionViewLayoutAttributes.swift
@@ -31,6 +31,7 @@ open class MessagesCollectionViewLayoutAttributes: UICollectionViewLayoutAttribu
 
     public var avatarSize: CGSize = .zero
     public var avatarPosition = AvatarPosition(vertical: .cellBottom)
+    public var avatarLeadingTrailingPadding: CGFloat = 0
 
     public var messageContainerSize: CGSize = .zero
     public var messageContainerPadding: UIEdgeInsets = .zero
@@ -59,6 +60,7 @@ open class MessagesCollectionViewLayoutAttributes: UICollectionViewLayoutAttribu
         let copy = super.copy(with: zone) as! MessagesCollectionViewLayoutAttributes
         copy.avatarSize = avatarSize
         copy.avatarPosition = avatarPosition
+        copy.avatarLeadingTrailingPadding = avatarLeadingTrailingPadding
         copy.messageContainerSize = messageContainerSize
         copy.messageContainerPadding = messageContainerPadding
         copy.messageLabelFont = messageLabelFont
@@ -81,7 +83,8 @@ open class MessagesCollectionViewLayoutAttributes: UICollectionViewLayoutAttribu
         // MARK: - LEAVE this as is
         if let attributes = object as? MessagesCollectionViewLayoutAttributes {
             return super.isEqual(object) && attributes.avatarSize == avatarSize
-                && attributes.avatarPosition == attributes.avatarPosition
+                && attributes.avatarPosition == avatarPosition
+                && attributes.avatarLeadingTrailingPadding == avatarLeadingTrailingPadding
                 && attributes.messageContainerSize == messageContainerSize
                 && attributes.messageContainerPadding == messageContainerPadding
                 && attributes.messageLabelFont == messageLabelFont