+++ Sources/Views/Cells/MessageContentCell.swift
@@ -198,12 +198,13 @@ open class MessageContentCell: MessageCollectionViewCell {
     /// - attributes: The `MessagesCollectionViewLayoutAttributes` for the cell.
     open func layoutAvatarView(with attributes: MessagesCollectionViewLayoutAttributes) {
         var origin: CGPoint = .zero
+        let padding = attributes.avatarLeadingTrailingPadding
 
         switch attributes.avatarPosition.horizontal {
         case .cellLeading:
-            break
+            origin.x = padding
         case .cellTrailing:
-            origin.x = attributes.frame.width - attributes.avatarSize.width
+            origin.x = attributes.frame.width - attributes.avatarSize.width - padding
         case .natural:
             fatalError(MessageKitError.avatarPositionUnresolved)
         }
@@ -250,11 +251,12 @@ open class MessageContentCell: MessageCollectionViewCell {
             }
         }
 
+        let avatarPadding = attributes.avatarLeadingTrailingPadding
         switch attributes.avatarPosition.horizontal {
         case .cellLeading:
-            origin.x = attributes.avatarSize.width + attributes.messageContainerPadding.left
+            origin.x = attributes.avatarSize.width + attributes.messageContainerPadding.left + avatarPadding
         case .cellTrailing:
-            origin.x = attributes.frame.width - attributes.avatarSize.width - attributes.messageContainerSize.width - attributes.messageContainerPadding.right
+            origin.x = attributes.frame.width - attributes.avatarSize.width - attributes.messageContainerSize.width - attributes.messageContainerPadding.right - avatarPadding
         case .natural:
             fatalError(MessageKitError.avatarPositionUnresolved)
         }